refactor: 서비스 패키지 및 클래스 이름 정리
