+++ services/mountain-service/src/main/java/com/deungsanlog/mountain/MountainServiceApplication.java
@@ -1,4 +1,4 @@
-package com.example.test2;
+package com.deungsanlog.mountain;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
 @SpringBootApplication
 @EnableDiscoveryClient
-public class Test2Application {
+public class MountainServiceApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(Test2Application.class, args);
+        SpringApplication.run(MountainServiceApplication.class, args);
     }
 
 }
