+++ services/mountain-service/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=ms2
+spring.application.name=mountain-service
 spring.profiles.active=dev
 spring.profiles.include=keys
