+++ services/mountain-service/src/test/java/com/deungsanlog/mountain/MountainServiceApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.test1;
+package com.deungsanlog.mountain;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class Test1ApplicationTests {
+class MountainServiceApplicationTests {
 
     @Test
     void contextLoads() {
