+++ services/user-service/src/main/java/com/deungsanlog/user/controller/MainController.java
@@ -1,4 +1,4 @@
-package com.example.test1.controller;
+package com.deungsanlog.user.controller;
 
 
 import org.springframework.stereotype.Controller;
