+++ services/user-service/src/main/resources/application.properties
@@ -1,3 +1,3 @@
-spring.application.name=ms1
+spring.application.name=user-service
 spring.profiles.active=dev
 spring.profiles.include=keys
